docs(): refine doc
Refs BRICK_STORE-2947,BRICK_STORE-2945
