+++ bricks/forms/src/interfaces/index.ts
@@ -4,3 +4,5 @@ export * from "./upload-files-v2";
 export * from "./general-date-picker";
 export * from "./general-cascader";
 export * from "./dynamic-form-item-v2";
+export * from "./upload-img";
+export * from "./user-or-user-group-select";
