+++ bricks/forms/src/interfaces/upload-img.ts
@@ -1,0 +1,4 @@
+export interface UploadImgValue {
+  text?: string;
+  images?: any[];
+}
