+++ bricks/forms/src/interfaces/user-or-user-group-select.ts
@@ -1,0 +1,4 @@
+export type UserOrUserGroupSelectValue = {
+  selectedUser: string[];
+  selectedUserGroup: string[];
+};
